REFACTOR: #13 리뷰 반영
